xseg_posix: Simplify posix_allocate function
* Use ftruncate syscall to expand the shared memory segment
  instead of lseek and write.

* Return an int instead of long in posix_allocate and
  posix_deallocate functions, the long value was never
  used for it's real value anyway.

+++ src/include/xseg/xseg.h
@@ -118,8 +118,8 @@ struct xseg_port;
 
 struct xseg_operations {
 	void  (*mfree)(void *mem);
-	long  (*allocate)(const char *name, uint64_t size);
-	long  (*deallocate)(const char *name);
+	int   (*allocate)(const char *name, uint64_t size);
+	int   (*deallocate)(const char *name);
 	void *(*map)(const char *name, uint64_t size, struct xseg *seg);
 	void  (*unmap)(void *xseg, uint64_t size);
 };
